Responsive checkout, profile
working on breaking points for checkout, checkout shipping, checkout success and profile picture. It is a bit messy but trying to keep the styling cohesive as possible, for future projects it will be better creating a cleaner base for the css and how elements interact with eachother while making things responsive.